composable gluon network
